[nextjs-api] api/view/gpx/files を express.js と共通化し next.js に実装
